+++ src/main/java/com/chaosbuffalo/mkultra/core/abilities/NotoriousDOT.java
@@ -45,7 +45,7 @@ public class NotoriousDOT extends PlayerToggleGroupAbility {
     }
 
     @Override
-    public int getManaCost(int currentRank) {
+    public float getManaCost(int currentRank) {
         return currentRank;
     }
